fix(tooling): scope the root TypeScript project to stop the tsserver OOM loop
The root tsconfig.json had no `include`, so with `allowJs: true` it globbed the
whole monorepo into one program: 9854 files, ~4.3 GB peak heap, 79s per check.
VS Code runs tsserver with --max-old-space-size=3072, so the root project could
never fit — it OOM'd and respawned continuously, leaving the editor stuck on
"Initializing 'tsconfig.json'". After scoping: 2238 files, 0.56 GB, 2.6s.

The five packages that had no `include` of their own gain `["src"]` so they stop
inheriting the root's root-relative one. No emit change — none of them has a .ts
outside src/, and each already set `rootDir: ./src`.

Scoping the root project also exposed that the GraphQLModules.GlobalContext
augmentation lived inside modules-app.ts, so it only applied when that entry
file was in the program. Moved to a standalone .d.ts with an identical type
surface.
