+++ packages/israeli-vat-scraper/tsconfig.json
@@ -5,5 +5,6 @@
     "outDir": "dist",
     "rootDir": "./src"
   },
+  "include": ["src"],
   "exclude": ["**/node_modules", "**/dist"]
 }
