+++ packages/pcn874-generator/tsconfig.json
@@ -4,5 +4,6 @@
     "outDir": "dist",
     "rootDir": "./src"
   },
+  "include": ["src"],
   "exclude": ["**/node_modules", "**/dist"]
 }
