+++ packages/server/src/modules-app.ts
@@ -46,22 +46,10 @@ import { sortCodesModule } from './modules/sort-codes/index.js';
 import { tagsModule } from './modules/tags/index.js';
 import { transactionsModule } from './modules/transactions/index.js';
 import { vatModule } from './modules/vat/index.js';
-import type { RawAuth } from './plugins/auth-plugin.js';
 import { ENVIRONMENT, RAW_AUTH } from './shared/tokens.js';
 import type { Environment } from './shared/types/index.js';
 
 const { Pool } = pg;
-
-declare global {
-  // eslint-disable-next-line @typescript-eslint/no-namespace
-  namespace GraphQLModules {
-    interface GlobalContext {
-      env: Environment;
-      rawAuth: RawAuth;
-      dbClientsToDispose?: { dispose: () => Promise<void> }[];
-    }
-  }
-}
 
 export async function createGraphQLApp(env: Environment, pool: pg.Pool) {
   const application = createApplication({
